+++ lucene/src/test/org/apache/lucene/search/similarities/Normalization.java
@@ -29,7 +29,7 @@ import org.apache.lucene.search.Explanation;
 public abstract class Normalization {
   /** Returns the normalized term frequency.
    * @param len the field length. */
-  public abstract float tfn(EasyStats stats, float tf, int len);
+  public abstract float tfn(EasyStats stats, float tf, float len);
   
   /** Returns an explanation for the normalized term frequency.
    * <p>The default normalization methods use the field length of the document
@@ -37,7 +37,7 @@ public abstract class Normalization {
    * This method provides a generic explanation for such methods.
    * Subclasses that use other statistics must override this method.</p>
    */
-  public Explanation explain(EasyStats stats, float tf, int len) {
+  public Explanation explain(EasyStats stats, float tf, float len) {
     Explanation result = new Explanation();
     result.setDescription(getClass().getSimpleName() + ", computed from: ");
     result.setValue(tfn(stats, tf, len));
@@ -51,12 +51,12 @@ public abstract class Normalization {
   /** Implementation used when there is no normalization. */
   public static final class NoNormalization extends Normalization {
     @Override
-    public final float tfn(EasyStats stats, float tf, int len) {
+    public final float tfn(EasyStats stats, float tf, float len) {
       return tf;
     }
 
     @Override
-    public final Explanation explain(EasyStats stats, float tf, int len) {
+    public final Explanation explain(EasyStats stats, float tf, float len) {
       return new Explanation(1, "no normalization");
     }
